backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.4
django-debug-toolbar==3.4.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
gunicorn==20.1.0
Markdown==3.3.7
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==6.1.0
